Repackage poly into subpackages. (#174)
* renamed hash package to seqhash to avoid collisions with stdlib.

* simplified read, write, parse, build names for all io packages.

* fixed package paths and removed depped test from hash package.

* fixed testing path.

* fixed named stutter for random.ProteinSequence.

* fixed spacing issue in random.go comment.

* made parseLocation private.

* renamed CodonTable to Table.

* renamed CloneSequence to Vector.

* renamed Vector to Part in clone.go.

* renamed BuildGbk to genbank.Build.

* removed stale tutorials in anticipation for new ones that will also function as tests.

* repackaged reverse into transform package itself and made accompanying reverse and complement functions and tests.

* changed json package to polyjson to avoid naming collisions with stdlib.

* added Example test for GetCodingRegions.

* added example test for Sequence.AddFeature.

* deprecated redundant .GetSequence() Sequence struct method.

* Fixed example test names and attained 100% test coverage for poly.go.

* Add part doc
