+++ .gitignore
@@ -2,3 +2,5 @@ data/genbank
 public
 node_modules
 *.DS_Store
+*.scratch
+*.vscode
