+++ io/fasta/data/uniprot_1mb_test.fasta.gz
