+++ io/uniprot/data/uniprot_sprot_mini.xml.gz
