+++ io/uniprot/uniprot_test.go
