+++ poly/commands_test.go
@@ -8,6 +8,8 @@ import (
 	"testing"
 
 	"github.com/TimothyStiles/poly"
+	"github.com/TimothyStiles/poly/io/genbank"
+	"github.com/TimothyStiles/poly/io/polyjson"
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
 )
@@ -58,7 +60,7 @@ func TestConvertPipe(t *testing.T) {
 
 		// getting test sequence from non-pipe io to compare against io to stdout
 		baseTestSequence := parseExt(match)
-		pipeOutputTestSequence := poly.ParseJSON(writeBuffer.Bytes())
+		pipeOutputTestSequence := polyjson.Parse(writeBuffer.Bytes())
 
 		if diff := cmp.Diff(baseTestSequence, pipeOutputTestSequence, cmpopts.IgnoreFields(poly.Feature{}, "ParentSequence")); diff != "" {
 			t.Errorf(" mismatch converting from %q to json (-want +got):\n%s", extension, diff)
@@ -169,8 +171,8 @@ func TestConvertFile(t *testing.T) {
 		t.Fatalf("Run error: %s", err)
 	}
 
-	puc19InputTestSequence := poly.ReadGbk("../data/puc19.gbk")
-	puc19OutputTestSequence := poly.ReadJSON("../data/puc19.json")
+	puc19InputTestSequence := genbank.Read("../data/puc19.gbk")
+	puc19OutputTestSequence := polyjson.Read("../data/puc19.json")
 
 	//clearing test data.
 	os.Remove("../data/puc19.json")
@@ -180,8 +182,8 @@ func TestConvertFile(t *testing.T) {
 		t.Errorf(" mismatch from concurrent gbk input test (-want +got):\n%s", diff)
 	}
 
-	t4InputTestSequence := poly.ReadGbk("../data/t4_intron.gb")
-	t4OutputTestSequence := poly.ReadJSON("../data/t4_intron.json")
+	t4InputTestSequence := genbank.Read("../data/t4_intron.gb")
+	t4OutputTestSequence := polyjson.Read("../data/t4_intron.json")
 
 	// clearing test data.
 	os.Remove("../data/t4_intron.json")
