+++ primers/primers.go
@@ -1,9 +1,11 @@
-package poly
+package primers
 
 import (
 	"bytes"
 	"math"
 	"strings"
+
+	"github.com/TimothyStiles/poly/transform"
 )
 
 // For reference: https://www.sigmaaldrich.com/technical-documents/articles/biology/oligos-melting-temp.html
@@ -56,7 +58,7 @@ func SantaLucia(sequence string, primerConcentration, saltConcentration, magnesi
 	dH += initialThermodynamicPenalty.H
 	dS += initialThermodynamicPenalty.S
 	// apply symmetry penalty if sequence is self-complementary
-	if sequence == ReverseComplement(sequence) {
+	if sequence == transform.ReverseComplement(sequence) {
 		dH += symmetryThermodynamicPenalty.H
 		dS += symmetryThermodynamicPenalty.S
 		symmetryFactor = 1
@@ -254,7 +256,7 @@ func CreateBarcodesWithBannedSequences(length int, maxSubSequence int, bannedSeq
 				barcodeNum++
 			}
 			// Check reverse complement as well for the banned sequence
-			for strings.Contains(debruijn[start:end], ReverseComplement(bannedSequence)) {
+			for strings.Contains(debruijn[start:end], transform.ReverseComplement(bannedSequence)) {
 				if end+1 > len(debruijn) {
 					return barcodes
 				}
