+++ seqhash/seqhash_test.go
@@ -1,27 +1,18 @@
-package poly
+package seqhash
 
 import (
 	"bytes"
 	"fmt"
 	"testing"
 
+	"github.com/TimothyStiles/poly/io/genbank"
 	"github.com/sergi/go-diff/diffmatchpatch"
 )
 
 func ExampleHash() {
-	sequence := ReadGbk("data/puc19.gbk")
+	sequence := genbank.Read("../data/puc19.gbk")
 
 	seqhash, _ := Hash(sequence.Sequence, "DNA", true, true)
-	fmt.Println(seqhash)
-	// output: v1_DCD_4b0616d1b3fc632e42d78521deb38b44fba95cca9fde159e01cd567fa996ceb9
-}
-
-func ExampleSequence_Hash() {
-	sequence := ReadGbk("data/puc19.gbk")
-
-	// Seqhash assumes doubleStranded sequence and defaults to linear
-	// if sequence.Meta.Locus.Circular is not set
-	seqhash, _ := sequence.Hash()
 	fmt.Println(seqhash)
 	// output: v1_DCD_4b0616d1b3fc632e42d78521deb38b44fba95cca9fde159e01cd567fa996ceb9
 }
@@ -83,37 +74,8 @@ func TestHash(t *testing.T) {
 
 }
 
-func TestSequence_Hash(t *testing.T) {
-	// Test no MoleculeType
-	sequence := ReadGbk("data/puc19.gbk")
-	sequence.Meta.Locus.MoleculeType = ""
-	_, err := sequence.Hash()
-	if err == nil {
-		t.Errorf("Seqhash method should fail with no MoleculeType")
-	}
-	// Test RNA sequence feature
-	sequence.Meta.Locus.MoleculeType = "RNA"
-	seqhash, _ := sequence.Hash()
-	if seqhash != "v1_RCD_4b0616d1b3fc632e42d78521deb38b44fba95cca9fde159e01cd567fa996ceb9" {
-		t.Errorf("RNA hashing failed. Expected v1_RCD_4b0616d1b3fc632e42d78521deb38b44fba95cca9fde159e01cd567fa996ceb9, got: " + seqhash)
-	}
-	// Test bad MoleculeType
-	sequence.Meta.Locus.MoleculeType = "TNA"
-	_, err = sequence.Hash()
-	if err == nil {
-		t.Errorf("Seqhash method should fail with TNA (or other) MoleculeType")
-	}
-	// Test bad sequence
-	sequence.Meta.Locus.MoleculeType = "DNA"
-	sequence.Sequence = "gagatacctacagcgtgagctatgagaaagcgccacgcttcccgaagggagaaaggcggacaggtatccggtaagcggcagggtcggaacaggagagcgcacgagggagcttccagggggaaacgcctggtatctttatagtcctgtcgggtttcgccacctctgacttgagcgtcgatttttgtgatgctcgtcaggggggcggagcctatggaaaaacgccagcaacgcggcctttttacggttcctggccttttgctggccttttgctcacatgttctttcctgcgttatcccctgattctgtggataaccgtattaccgcctttgagtgagctgataccgctcgccgcagccgaacgaccgagcgcagcgagtcagtgagcgaggaagcggaagagcgcccaatacgcaaaccgcctctccccgcgcgttggccgattcattaatgcagctggcacgacaggtttcccgactggaaagcgggcagtgagcgcaacgcaattaatgtgagttagctcactcattaggcaccccaggctttacactttatgcttccggctcgtatgttgtgtggaattgtgagcggataacaatttcacacaggaaacagctatgaccatgattacgccaagcttgcatgcctgcaggtcgactctagaggatccccgggtaccgagctcgaattcactggccgtcgttttacaacgtcgtgactgggaaaaccctggcgttacccaacttaatcgccttgcagcacatccccctttcgccagctggcgtaatagcgaagaggcccgcaccgatcgcccttcccaacagttgcgcagcctgaatggcgaatggcgcctgatgcggtattttctccttacgcatctgtgcggtatttcacaccgcatatggtgcactctcagtacaatctgctctgatgccgcatagttaagccagccccgacacccgccaacacccgctgacgcgccctgacgggcttgtctgctcccggcatccgcttacagacaagctgtgaccgtctccgggagctgcatgtgtcagaggttttcaccgtcatcaccgaaacgcgcgagacgaaagggcctcgtgatacgcctatttttataggttaatgtcatgataataatggtttcttagacgtcaggtggcacttttcggggaaatgtgcgcggaacccctatttgtttatttttctaaatacattcaaatatgtatccgctcatgagacaataaccctgataaatgcttcaataatattgaaaaaggaagagtatgagtattcaacatttccgtgtcgcccttattcccttttttgcggcattttgccttcctgtttttgctcacccagaaacgctggtgaaagtaaaagatgctgaagatcagttgggtgcacgagtgggttacatcgaactggatctcaacagcggtaagatccttgagagttttcgccccgaagaacgttttccaatgatgagcacttttaaagttctgctatgtggcgcggtattatcccgtattgacgccgggcaagagcaactcggtcgccgcatacactattctcagaatgacttggttgagtactcaccagtcacagaaaagcatcttacggatggcatgacagtaagagaattatgcagtgctgccataaccatgagtgataacactgcggccaacttacttctgacaacgatcggaggaccgaaggagctaaccgcttttttgcacaacatgggggatcatgtaactcgccttgatcgttgggaaccggagctgaatgaagccataccaaacgacgagcgtgacaccacgatgcctgtagcaatggcaacaacgttgcgcaaactattaactggcgaactacttactctagcttcccggcaacaattaatagactggatggaggcggataaagttgcaggaccacttctgcgctcggcccttccggctggctggtttattgctgataaatctggagccggtgagcgtgggtctcgcggtatcattgcagcactggggccagatggtaagccctcccgtatcgtagttatctacacgacggggagtcaggcaactatggatgaacgaaatagacagatcgctgagataggtgcctcactgattaagcattggtaactgtcagaccaagtttactcatatatactttagattgatttaaaacttcatttttaatttaaaaggatctaggtgaagatcctttttgataatctcatgaccaaaatcccttaacgtgagttttcgttccactgagcgtcagaccccgtagaaaagatcaaaggatcttcttgagatcctttttttctgcgcgtaatctgctgcttgcaaacaaaaaaaccaccgctaccagcggtggtttgtttgccggatcaagagctaccaactctttttccgaaggtaactggcttcagcagagcgcagataccaaatactgttcttctagtgtagccgtagttaggccaccacttcaagaactctgtagcaccgcctacatacctcgctctgctaatcctgttaccagtggctgctgccagtggcgataagtcgtgtcttaccgggttggactcaagacgatagttaccggataaggcgcagcggtcgggctgaacggggggttcgtgcacacagcccagcttggagcgaacgacctacaccgaactx"
-	_, err = sequence.Hash()
-	if err == nil {
-		t.Errorf("Seqhash method should fail with X nucleotide")
-	}
-}
-
 func ExampleRotateSequence() {
-	sequence := ReadGbk("data/puc19.gbk")
+	sequence := genbank.Read("../data/puc19.gbk")
 	sequenceLength := len(sequence.Sequence)
 	testSequence := sequence.Sequence[sequenceLength/2:] + sequence.Sequence[0:sequenceLength/2]
 
@@ -122,7 +84,7 @@ func ExampleRotateSequence() {
 }
 
 func TestLeastRotation(t *testing.T) {
-	sequence := ReadGbk("data/puc19.gbk")
+	sequence := genbank.Read("../data/puc19.gbk")
 	var sequenceBuffer bytes.Buffer
 
 	sequenceBuffer.WriteString(sequence.Sequence)
